# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Management API",
        "description": "API para la gestión de órdenes de restaurante. Permite crear órdenes con sus ítems, listar órdenes activas, avanzar en su estado ('initiated' → 'sent' → 'delivered') y eliminarlas automáticamente cuando se entregan.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Listar todas las órdenes",
                "operationId": "getOrders",
                "responses": {
                    "200": {
                        "description": "Lista de órdenes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "client_name": {
                                                        "type": "string",
                                                        "example": "Oran Kiehn I"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "initiated"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-08-07T18:34:10Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-08-07T18:34:10Z"
                                                    },
                                                    "items": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "order_id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "description": {
                                                                    "type": "string",
                                                                    "example": "consequatur aut"
                                                                },
                                                                "quantity": {
                                                                    "type": "integer",
                                                                    "example": 5
                                                                },
                                                                "unit_price": {
                                                                    "type": "string",
                                                                    "example": "28.45"
                                                                },
                                                                "created_at": {
                                                                    "type": "string",
                                                                    "format": "date-time",
                                                                    "example": "2025-08-07T18:34:10Z"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string",
                                                                    "format": "date-time",
                                                                    "example": "2025-08-07T18:34:10Z"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowed"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Crear una nueva orden",
                "operationId": "storeOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_name",
                                    "items"
                                ],
                                "properties": {
                                    "client_name": {
                                        "type": "string",
                                        "example": "Carlos Gómez"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "description",
                                                "quantity",
                                                "unit_price"
                                            ],
                                            "properties": {
                                                "description": {
                                                    "type": "string",
                                                    "example": "Lomo saltado"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 60
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowed"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtener una orden por ID",
                "operationId": "getOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos de la orden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowed"
                    }
                }
            }
        },
        "/api/orders/{id}/advance": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Avanzar el estado de una orden",
                "operationId": "advanceOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orden actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Estado actualizado a En preparación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowed"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreOrderRequest": {
                "required": [
                    "client_name",
                    "items"
                ],
                "properties": {
                    "client_name": {
                        "type": "string",
                        "example": "Cliente recurrente"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "required": [
                                "description",
                                "quantity",
                                "unit_price"
                            ],
                            "properties": {
                                "description": {
                                    "type": "string",
                                    "maxLength": 500,
                                    "example": "items description 01"
                                },
                                "quantity": {
                                    "type": "integer",
                                    "minimum": 1,
                                    "example": 4
                                },
                                "unit_price": {
                                    "type": "number",
                                    "format": "float",
                                    "minimum": 0,
                                    "example": 2
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Order": {
                "required": [
                    "client_name"
                ],
                "properties": {
                    "client_name": {
                        "type": "string",
                        "example": "Cliente recurrente"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "required": [
                    "description",
                    "quantity",
                    "unit_price"
                ],
                "properties": {
                    "description": {
                        "type": "string",
                        "maxLength": 500,
                        "example": "items description 01"
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 4
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 2
                    }
                },
                "type": "object"
            },
            "OrderResponse": {
                "required": [
                    "order",
                    "total"
                ],
                "properties": {
                    "order": {
                        "$ref": "#/components/schemas/Order"
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 8
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "NotFound": {
                "description": "Recurso no encontrado",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Recurso no encontrado"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ValidationError": {
                "description": "Datos inválidos",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Datos inválidos."
                                },
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ServerError": {
                "description": "Error interno del servidor",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Error interno del servidor"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "MethodNotAllowed": {
                "description": "Método HTTP no permitido",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Método HTTP no permitido para esta ruta."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Gestión de órdenes del restaurante"
        }
    ]
}